Re-organised the live services
